Fase Final

+++ app/Http/Controllers/CobradorController.php
@@ -1,11 +1,19 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Cobrador;
+use App\User;
+use App\Http\Requests\CobradorRequest;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Hash;
+use Carbon\Carbon;
 
 class CobradorController extends Controller
 {
+
+    public function __construct(){
+        $this->middleware(['auth', 'role:SuperUser|Prestamista|Administrador']);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -13,7 +21,7 @@ class CobradorController extends Controller
      */
     public function index()
     {
-        $cobradores = Cobrador::all();
+        $cobradores = User::role('Cobrador')->get();
         return view('auth.cobradores.index', compact('cobradores'));
     }
 
@@ -34,26 +42,23 @@ class CobradorController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CobradorRequest $request)
     {
-        $request = $this->validate(request(),[
-            'nombre' => 'required|string',
-            'paterno' => 'required|string',
-            'materno' => 'required|string',
-            'direccion' => 'required|string',
-            'telefono' => 'required'
-        ]);
+        $data = $request->validated();
         try {
-            $cobrador = new Cobrador;
-            $cobrador->nombre = $request['nombre'];
-            $cobrador->paterno = $request['paterno'];
-            $cobrador->materno = $request['materno'];
-            $cobrador->direccion = $request['direccion'];
-            $cobrador->telefono = $request['telefono'];
+            $cobrador = new User;
+            $cobrador->username  = $data['username'];
+            $cobrador->email = $data['email'];
+            $cobrador->password = bcrypt($data['password']); 
+            $cobrador->nombre = $data['nombre'];
+            $cobrador->paterno = $data['paterno'];
+            $cobrador->materno = $data['materno'];
+            $cobrador->fecha_na = Carbon::parse($data['nacimiento']);
             $cobrador->save();
+            $cobrador->assignRole('Cobrador');
             return redirect()->route('cobradores.index')->with('success', 'El nuevo registro se ha guardado correctamente');
         } catch (Exception $e) {
-            
+    
         }
 
     }
@@ -77,8 +82,20 @@ class CobradorController extends Controller
      */
     public function edit($id)
     {
-        $cobrador = Cobrador::find($id);
-        return view('auth.cobradores.edit', compact('cobrador'));
+            try{ 
+            $cobrador = User::find($id);
+        
+            if($cobrador == "" || strval($cobrador->hasRole('Cobrador')) !=1){
+                throw new \Exception;
+            }
+          
+
+                return view('auth.cobradores.edit', compact('cobrador'));
+
+            } catch (\Exception $e) {
+                return abort('500');
+            }
+        
     }
 
     /**
@@ -90,18 +107,31 @@ class CobradorController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $cobrador = Cobrador::find($id);
+        $cobrador = User::find($id);
+            $data = $this->validate(Request(),[
+                'username' => 'required|string|unique:usuarios,username,'.$cobrador->id,
+                'email' => 'required|email|unique:usuarios,email,'.$cobrador->id,
+                'nombre' => 'required|string',
+                'paterno' => 'required|string',
+                'materno' => 'required|string',
+                'nacimiento' => 'required|date',
+            ]);
+        try{
 
-        $cobrador->nombre = $request->get('nombre');
-        $cobrador->paterno = $request->get('paterno');
-        $cobrador->materno = $request->get('materno');
-        $cobrador->direccion = $request->get('direccion');
-        $cobrador->telefono = $request->get('telefono');
-       
-
-        $cobrador->save();
-    
-        return redirect()->route('cobradores.edit', $id)->with('success','El registro se ha modificado correctamente');
+            if($cobrador == "" || strval($cobrador->hasRole('Cobrador')) !=1){
+                throw new \Exception;
+            }
+            $cobrador->username = $data['username'];
+            $cobrador->email = $data['email'];
+            $cobrador->nombre = $data['nombre'];
+            $cobrador->paterno = $data['paterno'];
+            $cobrador->materno = $data['materno'];
+            $cobrador->fecha_na = Carbon::parse($data['nacimiento']);
+            $cobrador->save();    
+                return redirect()->route('cobradores.edit', $id)->with('success','El registro se ha modificado correctamente');
+        } catch (\Exception $e) {
+            return abort('500');
+        }    
     }
 
     /**
@@ -112,8 +142,60 @@ class CobradorController extends Controller
      */
     public function destroy($id)
     {
-        $cobrador= Cobrador::find($id);
-        $cobrador->delete();
-         return redirect()->route('cobradores.index')->with('success', 'El registro se ha eliminado correctamente');
+        $cobrador= user::find($id);
+        try {
+            if($cobrador == "" || strval($cobrador->hasRole('Cobrador')) !=1){
+                throw new \Exception;
+            }
+            $cobrador->delete();
+            return redirect()->route('cobradores.index')->with('success', 'El registro se ha eliminado correctamente');
+        } catch (\Exception $th) {
+            return abort('500');
+        }
+
+       
+    }
+
+    public function updatePassword(Request $request, $id){
+        $cobrador = user::find($id);
+        $data = $this->validate(Request(),[
+            'current_password' => 'required|string',
+            'password' => 'required|string|min:8',
+            'confirm_password' => 'required|same:password',
+        ]);
+        try {
+            if($cobrador == "" || strval($cobrador->hasRole('Cobrador')) !=1){
+                throw new \Exception;
+            }
+                $user = User::find($id);
+                if(Hash::check($data['current_password'], $user->password )){
+                    $user->password = Hash::make($data['password']);
+                    $user->save();
+                    return redirect()->route('user.edit', $user->id)->with('success', 'La contraseña se ha actualizado');
+                }else{
+                    return redirect()->back()->withErrors(['La contraseña actual no coincide']);
+                }
+        } catch (Exception $e) {
+             return abort('500');
+        }
+        
+    }
+
+    public function cobradorShowAjax(Request $request){
+        if($request->ajax()){
+            $data = $this->validate(Request(),[
+                'query' => 'nullable|string'
+            ]);
+            $query = $data['query']; 
+            if ($query != '') {
+                $data = User::role('Cobrador')->where('nombre', 'like', $query.'%')
+                ->orWhere('paterno', 'like', $query.'%')
+                ->orWhere('materno', 'like', $query.'%')
+                ->orWhere('email', 'like', $query.'%')->get();
+            }else{
+                $data = User::role('Cobrador')->get();
+            }
+            return response()->json(['data' => $data]);
+        }
     }
 }
